courses-api-ddd/requirements.txt
flask
python-jose
flask-restx